+++ Tether/Tether.csproj
@@ -115,6 +115,7 @@
     <Compile Include="Checks\DriveInfoBasedDiskUsageCheck.cs" />
     <Compile Include="Checks\IISCheck.cs" />
     <Compile Include="Checks\IOCheck.cs" />
+    <Compile Include="MetricJsonConverter.cs" />
     <Compile Include="Metrics\DiskUsageMetricProvider.cs" />
     <Compile Include="Metrics\IOMetricProvider.cs" />
     <Compile Include="Checks\NetworkTrafficCheck.cs" />
